Se agregó funcionalidad cambiar notificación estado y se le agregó menu y app bar al módulo mi perfil
