+++ lib/src/routes/routes.dart
@@ -3,7 +3,7 @@ import 'package:salud_y_mas/src/pages/landing_page.dart';
 import 'package:salud_y_mas/src/pages/login_page.dart';
 import 'package:salud_y_mas/src/pages/notificaciones/notificaciones_page.dart';
 import 'package:salud_y_mas/src/pages/pantalla_inicio.dart';
-import 'package:salud_y_mas/src/pages/perfil/perfil_page.dart';
+import 'package:salud_y_mas/src/pages/perfil/perfilUsuario.dart';
 import 'package:salud_y_mas/src/pages/revistas/revistas_page.dart';
 
 Map<String, WidgetBuilder> getApplicationRoutes() {
@@ -13,6 +13,6 @@ Map<String, WidgetBuilder> getApplicationRoutes() {
     'notificaciones': (BuildContext context) => NotificacionesPage(),
     'inicio': (BuildContext context) => HomePage(),
     'revistas': (BuildContext context) => RevistasPage(),
-    'perfil': (BuildContext context) => PerfilPage()
+    'perfil': (BuildContext context) => PerfilUsuario('')
   };
 }
