+++ lib/src/widgtes/menu.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:salud_y_mas/preferences/preferences.dart';
-import 'package:salud_y_mas/src/models/notificaciones_model.dart';
 import 'package:salud_y_mas/src/pages/notificaciones/notificaciones_page.dart';
-import 'package:salud_y_mas/src/pages/perfilUsuario.dart';
+import 'package:salud_y_mas/src/pages/perfil/perfilUsuario.dart';
 
 import 'alerts.dart';
 
@@ -100,8 +99,8 @@ class _MenuPageState extends State<MenuPage> {
                         child: ListTile(
                       hoverColor: Colors.purple,
                       contentPadding: EdgeInsets.fromLTRB(10, 0, 10, 0),
-                          leading: Icon(Icons.notifications),
-                          title: Text(
+                      leading: Icon(Icons.notifications),
+                      title: Text(
                         "Notificaciones",
                         style: TextStyle(),
                       ),
